requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
pypdf
python-docx
httpx